ix for the product (sum (f(i), i, 1, inf), j, 1, inf) bug
o Fix for the product (sum (f(i), i, 1, inf), j, 1, inf) bug:

In mult-expr-infinities, I moved the call to csign to only those cases that needed the csign. I don't know why this fixed the bug, but it did.

I'm not sure this is an entirely satisfactory fix, so I made a source code comments about this bug.